Move macros to their own file
